# Fix: Implement adaptive calibration for accurate time budget utilization